requirements.txt
# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# Machine Learning
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
torch>=2.0.1
torchaudio>=2.0.2

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Deep Learning Utilities
transformers>=4.31.0
resemblyzer>=0.1.1

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.1

# Web Interface
streamlit>=1.25.0
fastapi>=0.101.1
uvicorn>=0.23.2
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.1
joblib>=1.3.2
pyyaml>=6.0.1
scipy>=1.11.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.1
